+++ src/parameters.ts
@@ -1,6 +1,6 @@
 import { Colors } from "discord.js";
 
-import { getVersion } from "./helpers/getVersion.js";
+import { getVersion } from "./utils/getVersion.js";
 
 /** The version number of this project. */
 export const version = getVersion();
