+++ src/utils/formatPrompt.ts
