+++ src/utils/getVersion.ts
