+++ src/utils/messageAnalysis.ts
